Raise error if shipment id not found, remove id dependency in insert

+++ backend/app/db.py
@@ -9,6 +9,9 @@ class Database:
         self.cursor = self.connection.cursor()
         self.create_table()
 
+    def __del__(self):
+        self.close()
+
     def create_table(self):
         self.cursor.execute('''
             CREATE TABLE IF NOT EXISTS shipments (
@@ -22,22 +25,18 @@ class Database:
         self.connection.commit()
     
     def insert_shipment(self, shipment: ShipmentCreate) -> int:
-        self.cursor.execute("SELECT MAX(id) FROM shipments")
-        result = self.cursor.fetchone()
-        id = result[0] + 1 if result[0] is not None else 1
         self.cursor.execute('''
-            INSERT INTO shipments 
-            VALUES (:id, :content, :weight, :destination, :status)
+            INSERT INTO shipments (content, weight, destination, status)
+            VALUES (:content, :weight, :destination, :status)
         ''',
             {
-                "id": id, 
-                 **shipment.model_dump(),
+                **shipment.model_dump(),
                 "status": "Placed"
             }
         )
         self.connection.commit()
 
-        return id
+        return self.cursor.lastrowid
 
     def get_shipment(self, id: int) -> dict[str, Any] | None:
         self.cursor.execute('''
@@ -65,9 +64,6 @@ class Database:
             }
         )
         self.connection.commit()
-
-        if self.get_shipment(id) is None:
-            raise ValueError(f"Shipment with id {id} not found.")
         
         updated_shipment = self.get_shipment(id)
         if updated_shipment is None:
@@ -75,13 +71,15 @@ class Database:
         
         return updated_shipment
 
-
     def delete_shipment(self, id: int) -> None:
+        if self.get_shipment(id) is None:
+            raise ValueError(f"Shipment with id {id} does not exist.")
+        
         self.cursor.execute('''
             DELETE FROM shipments
             WHERE id = ?
         ''', (id,))
         self.connection.commit()
-    
+
     def close(self):
         self.connection.close()
